Add example.ex
