+++ languages/elixir/exercises/concept/recursion/lib/recursion.ex
@@ -1,18 +1,21 @@
 defmodule BirdCount do
-  def today([]), do: nil
-  def today([head | _]), do: head
+  def today(list) do
+    "Please implement the today/1 function"
+  end
 
-  def increment_day_count([]), do: [1]
-  def increment_day_count([head | tail]), do: [head + 1 | tail]
+  def increment_day_count(list) do
+    "Please implement the increment_day_count/1 function"
+  end
 
-  def has_day_without_birds?([]), do: false
-  def has_day_without_birds?([0 | _]), do: true
-  def has_day_without_birds?([_ | tail]), do: has_day_without_birds?(tail)
+  def has_day_without_birds?(list) do
+    "Please implement the has_day_without_birds?/1 function"
+  end
 
-  def total([]), do: 0
-  def total([head | tail]), do: head + total(tail)
+  def total(list) do
+    "Please implement the total/1 function"
+  end
 
-  def busy_days([]), do: 0
-  def busy_days([head | tail]) when head >= 5, do: busy_days(tail) + 1
-  def busy_days([_ | tail]), do: busy_days(tail)
+  def busy_days(list) do
+    "Please implement the total/1 function"
+  end
 end
